% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_divergence.R
\name{calculate2dJsDivergence}
\alias{calculate2dJsDivergence}
\title{Jensen-Shannon divergence between two-dimensional maps.}
\usage{
calculate2dJsDivergence(x, y, n = 2^8)
}
\arguments{
\item{x, y}{Two-dimensional numeric matrices to compare.}

\item{n}{Number of grid points for KDE.}
}
\value{
A numeric for the JS divergence between the maps.
}
\description{
Calculate the Jensen-Shannon (JS) divergence between two two-dimensional maps.
These maps may be generated by a dimensionality reduction (DR) method; however
this is not a requirement.
}
\seealso{
\code{\link{calculateJsDivergence}}
}
