% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aof.R
\name{calculateAof}
\alias{calculateAof}
\title{Average Overlap Frequency (AOF).}
\usage{
calculateAof(x, pos_indices, neg_indices, width = 0.05, cofactor = NULL)
}
\arguments{
\item{x}{Numeric vector corresponding to one channel in cytometry data.}

\item{pos_indices}{Indices of cells positive for this channel.}

\item{neg_indices}{Indices of cells negative for this channel.}

\item{width}{Width of high threshold of negative population and low
threshold of positive population.}

\item{cofactor}{If supplied, data will be transformed using inverse
hyperbolic sin with given cofactor.}
}
\value{
The AOF between positive and negative populations for x.
}
\description{
Calculate the Average Overlap Frequency (AOF) for a single cytometry channel.
}
