% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aof.R
\name{calculateCytometryAof}
\alias{calculateCytometryAof}
\title{Cytometry calculation of Average Overlap Frequency (AOF).}
\usage{
calculateCytometryAof(fcs_data, y, channel_names = colnames(fcs_data),
  cofactor = NULL)
}
\arguments{
\item{fcs_data}{A numerical matrix (NxD) of acquired cytometry data. Each row
corresponds to a cell, each column to a channel.}

\item{channel_names}{A character vector which lists the columns (channels) in
fcs_data for which to calculate the AOF.}

\item{cofactor}{A numeric. If specified, fcs_data will be transformed using
inverse hyperbolic sin (arcsinh) with this cofactor.}

\item{pos_indices}{A list of}
}
\value{
A data frame with AOF values for each channel.
}
\description{
Calculate the Average Overlap Frequency (AOF) for a cytometry data matrix.
}
