% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_divergence.R
\name{calculateDrJsDivergence}
\alias{calculateDrJsDivergence}
\title{Jensen-Shannon divergence between dimensionality reduction maps.}
\usage{
calculateDrJsDivergence(x, y, n = 2^8)
}
\arguments{
\item{x, y}{Two-dimensional numeric matrices to compare.}

\item{n}{Number of grid points for KDE.}
}
\value{
A numeric for the JS divergence between the maps.
}
\description{
Calculate the Jensen-Shannon (JS) divergence between two two-dimensional maps
that were generated by a dimensionality reduction (DR) method.
}
\seealso{
\code{\link{calculateJsDivergence}}
}
