% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_divergence.R
\name{calculateJsDivergence}
\alias{calculateJsDivergence}
\title{Calculate the Jensen-Shannon divergence.}
\usage{
calculateJsDivergence(p, q)
}
\arguments{
\item{p, q}{Numerical vectors representing the probability distributions.}
}
\value{
A numeric for the JS divergence between the vectors.
}
\description{
Given two probability distributions, calculate the Jensen-Shannon (JS)
divergence between them.
}
\seealso{
\code{\link{calculateDrJsDivergence}}
}
