% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aof.R
\name{calculateMultiChannelAof}
\alias{calculateMultiChannelAof}
\title{Calculate Average Overlap Frequency (AOF) for multiple channels of interest.}
\usage{
calculateMultiChannelAof(channel_population_relationships_filepath,
  base_fcs_data_filepath, single_sample_labels, sample_id)
}
\arguments{
\item{channel_population_relationships_filepath}{Filepath of the channel 
population relationships csv file that outlines channels of interest as well 
as negative and positive populations for that channel.  Column names start with 
"channel", followed by cell population names (i.e. "b_cell", "t_cell", etc.). 
Cell values are TRUE, FALSE, or left empty.}

\item{base_fcs_data_filepath}{Filepath of the "base" FCS file. This file has
data for all cells from the sample, regardless of manual gating
population designation.}

\item{single_sample_labels}{A list illustrating the relationships between 
specific samples, cells, and population assignments designated via manual 
gating. This can be generated via the `generatePopulationAssignments` function.}
}
\value{
A data frame with AOF values for each channel of interest, as denoted
by the channel_population_relationships csv file.
}
\description{
Calculate the Average Overlap Frequency (AOF) for multiple cytometry channels.
}
