% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderImportKey}
\alias{debarcoderImportKey}
\title{Import a barcoding key.}
\usage{
debarcoderImportKey(filename)
}
\arguments{
\item{filename}{CSV file from which to import the key.}
}
\value{
List with barcoding key, barcoding channel names, and number of
positive channels for each code.
}
\description{
Import a barcoding key from a CSV file. The CSV header should be of the
format:
}
\details{
code,channel_name_1,channel_name_2,...

Channel names should be explicit (such as Pd102Di). Channel values should be
denoted as 1 (channel positive for code) or 0 (channel negative for code).
}
