% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderLabelEvents}
\alias{debarcoderLabelEvents}
\title{Label events.}
\usage{
debarcoderLabelEvents(exprs_list, key, unlabeled_label = "unlabeled")
}
\arguments{
\item{exprs_list}{A list with expression and sorted expression matrices.}

\item{key}{Barcoding key data frame.}

\item{unlabeled_label}{Label for events that don't match any key codes.}
}
\value{
A data frame with a row for every event and a single Label column.
}
\description{
For each event, label it with the barcoding code that matches that event's
top N highest intensity channels.
}
