% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderPlots}
\alias{debarcoderPlots}
\title{Debarcoder diagnostic plots.}
\usage{
debarcoderPlots(path_prefix, labels, exprs_list)
}
\arguments{
\item{path_prefix}{Either NULL (do not export figures) or prefix to path
where files will be exported.}

\item{labels}{Data frame with an event label column.}

\item{exprs_list}{A list with expression and sorted expression matrices.}
}
\value{
A list of ggplot objects corresponding to the figures.
}
\description{
Given an FCS file, generate several debarcoder diagnostic plots. These
includea separate scatter plot of Mahalanobis ratio versus barcoding
separation distance, Mahalanobis ratio distribution, and barcoding separation
distance distribution.
}
