% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderPrepareFcs}
\alias{debarcoderPrepareFcs}
\title{Prepare FCS data for debarcoding.}
\usage{
debarcoderPrepareFcs(fcs, key, cofactor = 10)
}
\arguments{
\item{fcs}{FCS flow frame.}

\item{key}{Barcoding key data frame.}

\item{cofactor}{Before scaling, data will be transformed using inverse
hyperbolic sin with given cofactor.}
}
\value{
Expression data after transformation, scaling, and matching, and an
additional copy after sorting.
}
\description{
Transform expression data using cofactor, scale to [0..1] range, match to
barcoding channels, and sort each event separately.
}
