% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_divergence.R
\name{generate2dJsDivergenceDataFrame}
\alias{generate2dJsDivergenceDataFrame}
\title{Jensen-Shannon divergence between multiple two-dimensional maps.}
\usage{
generate2dJsDivergenceDataFrame(source_filepaths,
  two_d_map_column_names = c("tSNE1", "tSNE2"), n = 2^8)
}
\arguments{
\item{source_filepaths}{A vector containing the file paths for the data of interest.}

\item{two_d_map_column_names}{A vector containing the column names of interest. Defaults to Cytobank's default tSNE column names.}

\item{n}{Number of grid points for KDE.}
}
\value{
A data frame that includes a numeric for the JS divergence between each 
possible combination of generated 2-D maps.
}
\description{
Calculate the Jensen-Shannon (JS) divergence between two or more two-dimensional 
maps and generate a data frame displaying the divergence value for each possible 
pair of two-dimensional maps. The two-dimensional maps are generated from 
the source files and columns of interest (two_d_map_column_names). The maps 
may be generated by a dimensionality reduction (DR) method; however this is 
not a requirement.
}
\seealso{
\code{\link{calculateJsDivergence}}
}
