% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aof.R
\name{generate_population_assignments}
\alias{generate_population_assignments}
\title{Generate Population Assignments for Manually Gated Data}
\usage{
generate_population_assignments(clustering_channels, manual_labeling_filepath,
  samples_filepath, data_dir)
}
\arguments{
\item{clustering_channels}{Vector containing clustering channels (ex: c("Er167Di_CCR7",
"Er168Di_CD3"))}

\item{manual_labeling_filepath}{Path to csv file with columns representing 
sample_id, base, and manually gated populations. Each row contains
a sample_id as well as the corresponding base FCS filename and FCS files for
manually gating populations. Note: The base FCS file contains data for all 
cells from the sample, regardless of manual gating population designation}

\item{samples_filepath}{Filepath of the samples csv file that outlines sample 
IDs and corresponding FCS file paths.}

\item{data_dir}{Path to directory containing all FCS files.}
}
\value{
A list illustrating the relationships between specific samples, cells,
and population assignments designated via manual gating.
}
\description{
Generate a list with keys representing sample IDs (ex: sample_1, sample_2).
Values are data frames with columns representing cell populations (ex: 
sample_1.basophil, sample_1.b_cell). Data frame cell values are booleans.
}
