% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aof.R
\name{greedyCytometryAof}
\alias{greedyCytometryAof}
\title{Greedy search for optimal Average Overlap Frequency (AOF) values.}
\usage{
greedyCytometryAof(fcs_data, y, channel_names = colnames(fcs_data),
  width = 0.05, cofactor = NULL, verbose = TRUE)
}
\arguments{
\item{fcs_data}{A numerical matrix (NxD) of acquired cytometry data. Each row
corresponds to a cell, each column to a channel.}

\item{y}{A factor vector of length N which includes the cluster assignment of
each cell.}

\item{channel_names}{A character vector which lists the columns (channels) in
fcs_data for which to calculate the AOF.}

\item{width}{Width of high threshold of negative populationa and low
threshold of positive population.}

\item{cofactor}{If supplied, data will be transformed using inverse
hyperbolic sin with given cofactor.}

\item{verbose}{Boolean. Should the function report progress to terminal.}
}
\value{
A data frame with AOF values for each channel.
}
\description{
Given a cytometry data matrix and its clustering scheme, use a greedy search
strategy to find a partition of clusters that minimizes the AOF for each
marker.
}
