% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_rename.R
\name{importChannelNames}
\alias{importChannelNames}
\title{Import channel names and descriptions.}
\usage{
importChannelNames(filenames, verbose = TRUE)
}
\arguments{
\item{filenames}{List of FCS file names.}

\item{verbose}{If TRUE, script will report progress to console.}
}
\value{
Data frame with unique channel names and descriptions, channel mass,
and whether that mass repeats across files with different name or
description.
}
\description{
Given a list of FCS file names, import channel names and descriptions across
all files.
}
