% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_rename.R
\name{renameFcsFileChannels}
\alias{renameFcsFileChannels}
\title{Rename channels in set of FCS files.}
\usage{
renameFcsFileChannels(dest_path, filenames, channels, suffix = "renamed",
  verbose = TRUE)
}
\arguments{
\item{dest_path}{Directory where new FCS files are exported.}

\item{filenames}{List of FCS file names.}

\item{channels}{Channel rename data frame.}

\item{suffix}{New FCS file suffix.}

\item{verbose}{If TRUE, script will report progress to console.}
}
\description{
Given a list of FCS files and a channel rename data frame, import each of the
files, rename their channel names and descriptions, and export them as a new
file.
}
