% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_rename.R
\name{channelRename}
\alias{channelRename}
\title{FCS channel renaming.}
\usage{
channelRename(path, verbose = TRUE)
}
\arguments{
\item{path}{Directory where FCS files are located.}

\item{verbose}{If TRUE, script will report progress to console.}
}
\description{
A two-step utility for renaming FCS channel names and descriptions. This is
useful for situations where channels were misnamed across acquisitions (for
example, naming a channel CD8 in one sample and CD8a in another). See the
online documentation at "FCS Channel Renaming" for instructions.
}
