% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcode}
\alias{debarcode}
\title{Debarcode an FCS file.}
\usage{
debarcode(fcs_file_path, key_file_path, stern = "off", export_files = TRUE,
  export_figures = TRUE, verbose = TRUE)
}
\arguments{
\item{fcs_file_path}{FCS file path.}

\item{key_file_path}{CSV barcoding key file path.}

\item{stern}{Optional parameter for taking a sterner threshold. Lowers the
amount of debris and doublets but may remove more legitimate cells. Possible
options are "off" (default), "weak", and "strong".}

\item{export_files}{If TRUE, the function will export debarcoded FCS files.}

\item{export_figures}{If TRUE, the function will export diagnostic plots.}

\item{verbose}{If TRUE, the function will message the console with updates.}
}
\value{
A data frame with a row for every event. Columns are label, barcoding
separation distance, and Mahalanobis ratio.
}
\description{
Import FCS file and barcoding key and run it through the debarcoding
workflow: label events using Zunder et al., unlabel events according to
barcoding separation distance and Mahalanobis ratio heuristics, and export
each label to a separate file along with an accompanying diagnostic plot.
}
\details{
Please note that this function ends with a series of warnings due to
flowCore's write.FCS.
}
