% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderPlots}
\alias{debarcoderPlots}
\title{Debarcoder diagnostic plots.}
\usage{
debarcoderPlots(path_prefix, fcs, labels)
}
\arguments{
\item{path_prefix}{Either NULL (do not export figures) or prefix to path
where files will be expored. Each file will be named
[prefix_path].[label].jpg.}

\item{fcs}{FCS flow frame.}

\item{labels}{Data frame with an event label column.}
}
\value{
A list of ggplot objects corresponding to the figures.
}
\description{
Given an FCS file, generate a separate scatter plot of Mahalanobis ratio
versus barcoding separation distance.
}
