% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderUnlabelEvents}
\alias{debarcoderUnlabelEvents}
\title{Unlabel events.}
\usage{
debarcoderUnlabelEvents(exprs_list, labels, key, threshold_per = 0.99,
  unlabeled_label = "unlabeled")
}
\arguments{
\item{exprs_list}{A list with expression and sorted expression matrices.}

\item{labels}{Data frame with an event label column.}

\item{key}{Barcoding key data frame.}

\item{threshold_per}{Percentile for setting distance and ratio thresholds.}

\item{unlabeled_label}{Label for events that don't match any key codes.}
}
\value{
A data frame with a row for every event. Columns are label, barcoding
separation distance, and Mahalanobis ratio.
}
\description{
Identify a set of suspect events whose barcoding separation distance is below
a given threshold. Then, unlabel events if their Mahalanobis ratio is above
the threshold set by the suspect events.
}
\details{
Please consult the package documentation for a complete explanation of the
heuristics involved (TODO).
}
