% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debarcoder.R
\name{debarcoderUnlabelEvents}
\alias{debarcoderUnlabelEvents}
\title{Unlabel events.}
\usage{
debarcoderUnlabelEvents(exprs_list, labels, key, stern = "off",
  unlabeled_label = "unlabeled")
}
\arguments{
\item{exprs_list}{A list with expression and sorted expression matrices.}

\item{labels}{Data frame with an event label column.}

\item{key}{Barcoding key data frame.}

\item{stern}{Optional parameter for taking a sterner threshold. Lowers the
amount of debris and doublets but may remove more legitimate cells. Possible
options are "off" (default), "weak", and "strong".}

\item{unlabeled_label}{Label for events that don't match any key codes.}
}
\value{
A data frame with a row for every event. Columns are label, barcoding
separation distance, Mahalanobis ratio, and Mahalanobis distance.
}
\description{
Unlabel events if their Mahalanobis ratio is below the valley in the bimodal
Mahalanobis ratio distribution.
}
